+++ gulp/server.js
@@ -40,7 +40,7 @@ module.exports = function(options) {
     selector: '[ng-app]'// Only needed for angular apps
   }));
 
-  gulp.task('serve', ['watch'], function () {
+  gulp.task('serve', ['watch', 'prepare-translations'], function () {
     browserSyncInit([options.tmp + '/serve', options.src, options.src + '/**/*.js', options.lang]);
   });